Replace doubles with floats and remove draw_line function

+++ Structures/Camera.h
@@ -1,7 +1,7 @@
 #pragma once
 
 struct Camera {
-    double n, f; // near and far planes distance
-    double l, r; // near plane left and right edges coordinates
-    double b, t; // near plane bottom and top edges coordinates
+    float n, f; // near and far planes distance
+    float l, r; // near plane left and right edges coordinates
+    float b, t; // near plane bottom and top edges coordinates
 };
